// node-red-contrib-humaine/haic_client/schemas/EnvMeta.ts
// Generated by ts-to-zod
import { z } from "zod";

/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export const envMetaSchema = z.object({
    id: z.string(),
    name: z.string(),
    sim_id: z.string().optional().nullable(),
    domain: z.string().optional().nullable(),
    task: z.string().optional().nullable(),
    version: z.string().optional().nullable(),
    updated_at: z.string().optional().nullable()
});